Refactor total users kpi to fetch data with useSearchStrategy
