+++ x-pack/plugins/security_solution/common/search_strategy/security_solution/users/index.ts
@@ -5,7 +5,11 @@
  * 2.0.
  */
 
+import { TotalUsersKpiStrategyResponse } from './kpi/total_users';
+
 export enum UsersQueries {
   details = 'userDetails',
   kpiTotalUsers = 'usersKpiTotalUsers',
 }
+
+export type UserskKpiStrategyResponse = Omit<TotalUsersKpiStrategyResponse, 'rawResponse'>;
