--- x-pack/plugins/security_solution/public/users/containers/kpis/total_users/translations.ts
@@ -1,22 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { i18n } from '@kbn/i18n';
-
-export const ERROR_USERS_KPI = i18n.translate(
-  'xpack.securitySolution.hostsKpiHosts.errorSearchDescription',
-  {
-    defaultMessage: `An error has occurred on total users kpi search`,
-  }
-);
-
-export const FAIL_USERS_KPI = i18n.translate(
-  'xpack.securitySolution.hostsKpiHosts.failSearchDescription',
-  {
-    defaultMessage: `Failed to run search on total users kpi`,
-  }
-);
